fix: Press Ctrl only once while using key combinations.
The original reason why native JavaScript clicks were used in the
first place was the strange behavior of the clicks in GlobalSearchTest
test. It turned out that the clicks were performed with Ctrl key
pressed. Such a combination Ctrl + click opens the link under the
button in a new browser tab rather then clicks on the element.