Issue #291 : Remove SCRIPTS_DIR from nextflow docker.
Environment variable SCRIPTS_DIR was specified in nextflow image to
a shared directory. The problem is that shared directory cleans up
when shared file system services is configured. It led to a missing
pipeline repository files.

Also commit specifies main grid engine queue in the default nextflow
config for the docker image.