Use requirements.txt in setup.py of CP CLI. (#431)
Content of requirements.txt file is almost a duplicate of setuptools
requirements list. The inconsistency of the two configurations can be
pretty harmful as long as both configurations are expected to be
the same.

The commit forces setuptools to use requirements.txt file for a list
of package requirements. This way requirements.txt is packaged along
with sources in the pipe-cli package.