gui_issue-#794_error while generating .docx document was fixed.
Because of errors that we get in the form of a Blob, which parsed and appended to the new '.docx' document as content:
Added check if response data is Blob and has type field of 'application/json'
If so, we validate Blob object for errors in status field and if there is no errors, document proceed to generate.
Otherwise, if response is no a blob or don't have field 'application/json' - validation are skipped.